requirements.txt
asgiref==3.6.0
cffi==1.15.1
Django==4.1.5
django-environ==0.9.0
django-hosts==5.2
django-ptrack==3.0.0
djongo==1.3.6
dnspython==2.3.0
Pillow==9.4.0
pycparser==2.21
pymongo==3.12.3
PyNaCl==1.5.0
pytz==2022.7.1
sqlparse==0.2.4
tzdata==2022.7
